converted rcd charges values to float, made pipes a half a charge and, made rpd capacity 45.

+++ Content.Shared/RCD/Systems/RCDSystem.cs
@@ -604,14 +604,14 @@ public sealed partial class RCDDoAfterEvent : DoAfterEvent
     public ProtoId<RCDPrototype> StartingProtoId { get; private set; }
 
     [DataField]
-    public int Cost { get; private set; } = 1;
+    public float Cost { get; private set; } = 1;
 
     [DataField("fx")]
     public NetEntity? Effect { get; private set; }
 
     private RCDDoAfterEvent() { }
 
-    public RCDDoAfterEvent(NetCoordinates location, Direction direction, ProtoId<RCDPrototype> startingProtoId, int cost, NetEntity? effect = null)
+    public RCDDoAfterEvent(NetCoordinates location, Direction direction, ProtoId<RCDPrototype> startingProtoId, float cost, NetEntity? effect = null)
     {
         Location = location;
         Direction = direction;
